Moved the actions file prior to breaking it into its sub-parts
