+++ kakao/src/main/kotlin/com/agoda/kakao/Views.kt
@@ -14,6 +14,7 @@ import android.support.test.espresso.web.sugar.Web
 import android.support.v7.widget.AppCompatButton
 import android.support.v7.widget.AppCompatTextView
 import android.view.View
+import com.agoda.kakao.actions.*
 import com.agoda.kakao.builders.DataBuilder
 import com.agoda.kakao.builders.ViewBuilder
 import com.agoda.kakao.builders.WebElementBuilder
